webserver/requirements.txt
flask>=3.0.0
flask-socketio>=5.3.0
python-dotenv>=1.0.0
numpy>=1.24.0